cherrobot/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "chatbot_pack", from = "packages/chatbot_pack" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = ">=0.0.30"}
pypdf = "^5.1.0"
bs4 = "^0.0.2"
langchain-unstructured = "^0.1.5"
langchainhub = "^0.1.21"
langgraph = "^0.2.45"
langsmith = "^0.1.142"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
load-dotenv = "^0.1.0"
streamlit = "^1.40.1"
langchain-groq = "^0.2.1"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
